backend/requirement.txt
Flask
Pillow
tensorflow
numpy
androguard 
pandas 
Flask-Cors